Organize candidate evaluation and cost model modules
Place the model-facing Candidate, Cost, CostModel, and SizeCost under the cost module. Split Cost and CostModel into cost/model.rs, keep the size submodule private, and re-export every public cost-model type through cost::.

Rename the public estimate module to evaluation because it now defines scheme evaluation states, deferred resolution, and candidate evidence rather than selector scoring. Update every scheme implementation and internal link to use the new path.

Each concept now has one public path, and the source layout reflects the candidate-to-cost-model boundary.
